+++ packages/scss/demo/demo.js
@@ -6,6 +6,31 @@ function reloadAnimation(elem) {
 	setTimeout(function(){ card.classList.add(animationClass)},0);
 }
 
+
+/* TEXTFIELD INLINE VALIDATION */
+function validationFocus(elem) {
+	var textfield = elem.parentElement;
+	textfield.classList.remove('is-valid');
+	textfield.classList.add('is-loading');
+	setTimeout(function(){
+		textfield.classList.remove('is-loading');
+		textfield.classList.add('is-valid');
+	}, 1500);
+}
+
+function unvalidationFocus(elem) {
+	var textfield = elem.parentElement;
+	textfield.classList.remove('is-invalid');
+	textfield.classList.add('is-loading');
+	setTimeout(function(){
+		textfield.classList.remove('is-loading');
+		textfield.classList.add('is-invalid');
+	}, 1500);
+}
+
+
+
+
 /* ICONS */
 function generateIconDemo() {
 	var icons = {"icons": [